change format multer

+++ controllers/Product.js
@@ -26,6 +26,23 @@ export const createProduct = wrapAsync(async (req, res) => {
     });
 });
 
+export const updateProduct = wrapAsync(async (req, res) => {
+    const {
+        title, description, price,
+        ratting, stock, weight, category } = req.body;
+    const { id } = req.params;
+
+    const product = await Product.findByIdAndUpdate(id, {
+        title, description, price,
+        ratting, stock, weight, category
+    }, { new: true });
+
+    res.status(200).json({
+        status: 'success',
+        product
+    });
+});
+
 export const getAllProduct = wrapAsync(async (req, res) => {
     const product = await Product.find({});
 
@@ -88,8 +105,9 @@ export const searchProduct = wrapAsync(async (req, res) => {
 
 export const deleteProduct = wrapAsync(async (req, res) => {
     const { id } = req.params;
+    const userId = req.session.userId;
 
-    const product = await Product.findOneAndDelete({ _id: id });
+    const product = await Product.findOneAndDelete({ $and: [{ userId }, { _id: id }] });
 
     if (!product) {
         return res.status(404).json({
